add checker framework and unit tests github action
