+++ .github/workflows/unit-tests.yml
@@ -1,0 +1,45 @@
+name: Unit Tests
+
+on:
+  workflow_dispatch:
+  pull_request:
+    branches:
+      - 'master'
+      - 'develop'
+  push:
+    branches:
+      - 'master'
+      - 'develop'
+
+jobs:
+  tests:
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Set up JDK
+        uses: actions/setup-java@v3
+        with:
+          java-version: '20'
+          distribution: 'liberica'
+          cache: maven
+
+      - name: Maven Tests
+        run: mvn --batch-mode clean test
+
+      - name: Test Coverage
+        uses: codecov/codecov-action@v3
+
+      - name: SonarCloud Analyze
+        run: >
+          mvn --batch-mode sonar:sonar
+          -Dsonar.projectKey=spacious-team_investbook
+          -Dsonar.organization=spacious-team
+          -Dsonar.host.url=https://sonarcloud.io
+          -Dsonar.login=$SONAR_TOKEN
+          -Dsonar.coverage.jacoco.xmlReportPaths=./target/site/jacoco/jacoco.xml
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
