+++ src/test/java/ru/investbook/InvestbookApplicationTest.java
@@ -29,11 +29,11 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ApplicationContext;
 import org.springframework.test.web.servlet.MockMvc;
 
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.testng.Assert.assertNotNull;
 
 @SpringBootTest(classes = InvestbookApplication.class)
 @AutoConfigureMockMvc
@@ -65,5 +65,4 @@ class InvestbookApplicationTest {
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$." + jsonPath).value(expected));
     }
-
 }
