+++ src/test/java/ru/investbook/parser/psb/CashFlowTableTest.java
@@ -18,34 +18,39 @@
 
 package ru.investbook.parser.psb;
 
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mock;
-import org.spacious_team.broker.pojo.SecurityEventCashFlow;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Ignore;
-import org.testng.annotations.Test;
+import org.spacious_team.broker.pojo.EventCashFlow;
 import ru.investbook.parser.SecurityRegistrar;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.List;
 
-import static org.testng.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@Ignore
-public class DividendTableTest {
+@Disabled
+public class CashFlowTableTest {
 
     @Mock
     SecurityRegistrar securityRegistrar;
 
-    @DataProvider(name = "isin")
-    Object[][] getData() {
-        return new Object[][] {{"E:\\1.xlsx", "RU000A0JRKT8", "RU000A0JRKT8" }};
+    static Object[][] cash() {
+        return new Object[][]{{"E:\\1.xlsx", BigDecimal.valueOf(400 - 760.77)},
+                {"E:\\Налог.xlsx", BigDecimal.valueOf(-542.0)}};
     }
 
-    @Test(dataProvider = "isin")
-    void testIsin(String report, String firstIsin, String lastIsin) throws IOException {
-        PsbBrokerReport psbBrokerReport = new PsbBrokerReport(report, securityRegistrar);
-        List<SecurityEventCashFlow> data = new DividendTable(psbBrokerReport).getData();
-        assertEquals(data.get(0).getSecurity(), firstIsin);
-        assertEquals(data.get(data.size() - 1).getSecurity(), lastIsin);
+    @ParameterizedTest
+    @MethodSource("cash")
+    void testIsin(String reportFile, BigDecimal expectedCashIn) throws IOException {
+        PsbBrokerReport report = new PsbBrokerReport(reportFile, securityRegistrar);
+        List<EventCashFlow> data = new CashFlowTable(report).getData();
+        BigDecimal sum = BigDecimal.ZERO;
+        for (EventCashFlow r : data) {
+            sum = sum.add(r.getValue());
+        }
+        assertEquals(sum, expectedCashIn);
     }
 }
