+++ src/test/java/ru/investbook/parser/psb/DerivativeCashFlowTableTest.java
@@ -18,39 +18,38 @@
 
 package ru.investbook.parser.psb;
 
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mock;
-import org.spacious_team.broker.pojo.EventCashFlow;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Ignore;
-import org.testng.annotations.Test;
+import org.spacious_team.broker.pojo.SecurityEventCashFlow;
 import ru.investbook.parser.SecurityRegistrar;
 
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.util.List;
 
-import static org.testng.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@Ignore
-public class CashFlowTableTest {
+@Disabled
+public class DerivativeCashFlowTableTest {
 
     @Mock
     SecurityRegistrar securityRegistrar;
 
-    @DataProvider(name = "cash")
-    Object[][] getData() {
-        return new Object[][]{{"E:\\1.xlsx", BigDecimal.valueOf(400 - 760.77)},
-                {"E:\\Налог.xlsx", BigDecimal.valueOf(-542.0)}};
+    static Object[][] cashFlow() {
+        return new Object[][] {{"E:\\Исполнение фьючерса.xlsx", BigDecimal.valueOf(-733.0) }};
     }
 
-    @Test(dataProvider = "cash")
-    void testIsin(String reportFile, BigDecimal expectedCashIn) throws IOException {
-        PsbBrokerReport report = new PsbBrokerReport(reportFile, securityRegistrar);
-        List<EventCashFlow> data = new CashFlowTable(report).getData();
+    @ParameterizedTest
+    @MethodSource("cashFlow")
+    void testIsin(String report, BigDecimal expectedSum) throws IOException {
+        PsbBrokerReport psbBrokerReport = new PsbBrokerReport(report, securityRegistrar);
+        List<SecurityEventCashFlow> data = new DerivativeCashFlowTable(psbBrokerReport).getData();
         BigDecimal sum = BigDecimal.ZERO;
-        for (EventCashFlow r : data) {
+        for (SecurityEventCashFlow r : data) {
             sum = sum.add(r.getValue());
         }
-        assertEquals(sum, expectedCashIn);
+        assertEquals(sum, expectedSum);
     }
 }
