+++ src/test/java/ru/investbook/parser/psb/DividendTableTest.java
@@ -1,6 +1,6 @@
 /*
  * InvestBook
- * Copyright (C) 2020  Spacious Team <spacious-team@ya.ru>
+ * Copyright (C) 2023  Spacious Team <spacious-team@ya.ru>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -18,38 +18,34 @@
 
 package ru.investbook.parser.psb;
 
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mock;
 import org.spacious_team.broker.pojo.SecurityEventCashFlow;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Ignore;
-import org.testng.annotations.Test;
 import ru.investbook.parser.SecurityRegistrar;
 
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.util.List;
 
-import static org.testng.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@Ignore
-public class DerivativeCashFlowTableTest {
+@Disabled
+public class DividendTableTest {
 
     @Mock
     SecurityRegistrar securityRegistrar;
 
-    @DataProvider(name = "cash-flow")
-    Object[][] getData() {
-        return new Object[][] {{"E:\\Исполнение фьючерса.xlsx", BigDecimal.valueOf(-733.0) }};
+    static Object[][] getData() {
+        return new Object[][] {{"E:\\1.xlsx", "RU000A0JRKT8", "RU000A0JRKT8" }};
     }
 
-    @Test(dataProvider = "cash-flow")
-    void testIsin(String report, BigDecimal expectedSum) throws IOException {
+    @ParameterizedTest
+    @MethodSource("cashFisinlow")
+    void testIsin(String report, String firstIsin, String lastIsin) throws IOException {
         PsbBrokerReport psbBrokerReport = new PsbBrokerReport(report, securityRegistrar);
-        List<SecurityEventCashFlow> data = new DerivativeCashFlowTable(psbBrokerReport).getData();
-        BigDecimal sum = BigDecimal.ZERO;
-        for (SecurityEventCashFlow r : data) {
-            sum = sum.add(r.getValue());
-        }
-        assertEquals(sum, expectedSum);
+        List<SecurityEventCashFlow> data = new DividendTable(psbBrokerReport).getData();
+        assertEquals(data.get(0).getSecurity(), firstIsin);
+        assertEquals(data.get(data.size() - 1).getSecurity(), lastIsin);
     }
 }
