+++ src/test/java/ru/investbook/parser/psb/PortfolioCashTableTest.java
@@ -1,6 +1,6 @@
 /*
  * InvestBook
- * Copyright (C) 2020  Spacious Team <spacious-team@ya.ru>
+ * Copyright (C) 2023  Spacious Team <spacious-team@ya.ru>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -18,18 +18,18 @@
 
 package ru.investbook.parser.psb;
 
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mock;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Ignore;
-import org.testng.annotations.Test;
 import ru.investbook.parser.SecurityRegistrar;
 
 import java.io.IOException;
 import java.math.BigDecimal;
 
-import static org.testng.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@Ignore
+@Disabled
 public class PortfolioCashTableTest {
 
     @Mock
@@ -40,7 +40,8 @@ public class PortfolioCashTableTest {
         return new Object[][] {{"E:\\1.xlsx", BigDecimal.valueOf(350.37)}};
     }
 
-    @Test(dataProvider = "cash_in")
+    @ParameterizedTest
+    @MethodSource("cashIn")
     void testIsin(String report, BigDecimal expectedCash) throws IOException {
         PsbBrokerReport psbBrokerReport = new PsbBrokerReport(report, securityRegistrar);
         assertEquals(new CashTable(psbBrokerReport).getData().get(0).getValue(), expectedCash);
